Backend/requirements.txt
# ============================================================
# Web Framework
# ============================================================
fastapi==0.115.4
uvicorn[standard]==0.32.1
python-multipart==0.0.12
orjson==3.10.12

# ============================================================
# Database — PostgreSQL + pgvector
# ============================================================
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6

# ============================================================
# Cache & Task Queue
# ============================================================
redis==5.2.1
celery[redis]==5.4.0
flower==2.0.1
kombu==5.4.2

# ============================================================
# Authentication & Security
# ============================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==44.0.0

# ============================================================
# File Type Detection  (linux/docker: python-magic; installs libmagic)
# ============================================================
python-magic==0.4.27

# ============================================================
# Data Validation
# ============================================================
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]==2.10.3
email-validator==2.2.0

# ============================================================
# HTTP Client
# ============================================================
httpx==0.28.1
anyio==4.7.0
supabase

# ============================================================
# LLM Integration (OpenRouter via OpenAI-compatible SDK)
# ============================================================
openai==1.57.4

# ============================================================
# NLP & Embeddings
# ============================================================
sentence-transformers==3.3.1
spacy==3.8.3
nltk==3.9.1
rapidfuzz==3.11.0
# Run after install: python -m spacy download en_core_web_sm

# ============================================================
# Resume Parsing
# ============================================================
pdfplumber==0.11.4
pymupdf==1.25.1
python-docx==1.1.2
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0

# ============================================================
# Object Storage (S3 / MinIO)
# ============================================================
boto3==1.35.91
botocore==1.35.91

# ============================================================
# Structured Logging & Observability
# ============================================================
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0

# ============================================================
# Rate Limiting
# ============================================================
slowapi==0.1.9

# ============================================================
# Utilities
# ============================================================
python-dotenv==1.0.1
python-dateutil==2.9.0
Levenshtein==0.26.1
